Make tests green again
